still not working but now MkSerdes is implicitly available in Setup if there are Codec instances for the types, a registry and a mapping from schema tu parsed schema
